+++ src/main/java/org/mclods/simple_factory/products/impl/PepperoniPizza.java
@@ -1,6 +1,6 @@
-package org.mclods.simple_factory.product.impl;
+package org.mclods.simple_factory.products.impl;
 
-import org.mclods.simple_factory.product.Pizza;
+import org.mclods.simple_factory.products.Pizza;
 
 public class PepperoniPizza extends Pizza {
     public PepperoniPizza() {
